{{ cookiecutter.project_slug }}/requirements.txt
black==21.5b2
bcrypt==3.2.0
fastapi==0.66.0
gunicorn==20.1.0
isort==5.8.0
mypy==0.812
passlib==1.7.4
pip-chill==1.0.1
psycopg2==2.8.6
pyhumps==3.0.2
uvicorn==0.14.0
